Точка меняет число, немного увеличена точность числа после точки

+++ Calculator/Calculator/Abstractions/IInputService.cs
@@ -28,6 +28,6 @@
         /// <summary>
         /// Отдаёт число
         /// </summary>
-        double GetNumber();
+        decimal GetNumber();
     }
 }
